Partly done moving commands into logically separate files
