+++ package.json
@@ -7,7 +7,8 @@
   "scripts": {
     "build": "tsc --build",
     "clean": "tsc --build --clean",
-    "start": "node ./dist/index.js"
+    "start": "node ./dist/index.js",
+    "format": "prettier --write src/**/*.ts"
   },
   "author": "",
   "license": "ISC",
@@ -33,7 +34,7 @@
     "typescript": "^4.2.4"
   },
   "engines": {
-    "node": "18.2.0",
+    "node": "^18.2.0",
     "npm": "8.9.0"
   }
 }
