+++ src/commands/clan/index.ts
@@ -1,0 +1,6 @@
+import brains from "./brains";
+import skills from "./skills";
+import status from "./status";
+import whitelist from "./whitelist";
+
+export default { brains, skills, status, whitelist };
