+++ src/commands/index.ts
@@ -1,0 +1,5 @@
+import clan from "./clan";
+import kol from "./kol";
+import misc from "./misc";
+
+export default { ...clan, ...kol, ...misc };
