+++ src/commands/kol/index.ts
@@ -1,0 +1,9 @@
+import fairy from "./fairy";
+import item from "./item";
+import leaderboard from "./leaderboard";
+import leprechaun from "./leprechaun";
+import spade from "./spade";
+import stat from "./stat";
+import volleyball from "./volleyball";
+
+export default { fairy, item, leaderboard, leprechaun, spade, stat, volleyball };
