+++ src/commands/misc/index.ts
@@ -1,0 +1,7 @@
+import orb from "./orb";
+import prswelcome from "./prswelcome";
+import purge from "./purge";
+import remind from "./remind";
+import roll from "./roll";
+
+export default { orb, prswelcome, purge, remind, roll };
