+++ src/constants.ts
@@ -345,150 +345,6 @@ export const REVERSE_PACKAGES: Map<string, string> = new Map(
   reversed.concat(ghostlings).concat(foldables)
 );
 
-export const PATH_MAPPINGS: Map<string, number> = new Map([
-  ["clan", 8],
-  ["clandungeons", 8],
-  ["slimetube", 8],
-  ["slime", 8],
-  ["hobopolis", 8],
-  ["hobo", 8],
-  ["hobop", 8],
-  ["beeshateyou", 9],
-  ["bhy", 9],
-  ["bees", 9],
-  ["wayofthesurprisingfist", 10],
-  ["wotsf", 10],
-  ["fist", 10],
-  ["trendy", 11],
-  ["avatarofboris", 12],
-  ["aob", 12],
-  ["boris", 12],
-  ["bugbearinvasion", 13],
-  ["bugbear", 13],
-  ["zombieslayer", 14],
-  ["zombiemaster", 14],
-  ["zombie", 14],
-  ["classact", 15],
-  ["avatarofjarlsberg", 16],
-  ["jarlsberg", 16],
-  ["jarl", 16],
-  ["aoj", 16],
-  ["big", 17],
-  ["kolhs", 18],
-  ["hs", 18],
-  ["highschool", 18],
-  ["classactii", 19],
-  ["classact2", 19],
-  ["avatarofsneakypete", 20],
-  ["sneakypete", 20],
-  ["pete", 20],
-  ["aosp", 20],
-  ["slowandsteady", 21],
-  ["heavyrains", 22],
-  ["hr", 22],
-  ["picky", 23],
-  ["actuallyedtheundying", 24],
-  ["edtheundying", 24],
-  ["ed", 24],
-  ["undying", 24],
-  ["onecrazyrandomsummer", 25],
-  ["1crazyrandomsummer", 25],
-  ["ocrs", 25],
-  ["1crs", 25],
-  ["communityservice", 26],
-  ["cs", 26],
-  ["hccs", 26],
-  ["batfellow", 27],
-  ["avatarofwestofloathing", 28],
-  ["awol", 28],
-  ["thesource", 29],
-  ["source", 29],
-  ["nuclearautumn", 30],
-  ["na", 30],
-  ["gelatinousnoob", 31],
-  ["noob", 31],
-  ["gnoob", 31],
-  ["licensetoadventure", 32],
-  ["lta", 32],
-  ["liveascendrepeat", 33],
-  ["lar", 33],
-  ["pocketfamiliars", 34],
-  ["pokefam", 34],
-  ["glover", 35],
-  ["disguisesdelimit", 36],
-  ["dd", 36],
-  ["darkgyffte", 37],
-  ["dg", 37],
-  ["vampire", 37],
-  ["vampyre", 37],
-  ["twocrazyrandomsummer", 38],
-  ["2crazyrandomsummer", 38],
-  ["tcrs", 38],
-  ["2crs", 38],
-  ["kingdomofexploathing", 39],
-  ["exploathing", 39],
-  ["pathoftheplumber", 40],
-  ["plumber", 40],
-  ["mario", 40],
-  ["lowkeysummer", 41],
-  ["lks", 41],
-  ["greygoo", 42],
-  ["graygoo", 42],
-  ["goo", 42],
-  ["yourobot", 43],
-  ["robot", 43],
-  ["quantumterrarium", 44],
-  ["qt", 44],
-  ["wildfire", 45],
-  ["greyyou", 46],
-  ["grayyou", 46],
-  ["journeyman", 47],
-  ["dinosaur", 48],
-  ["fallofthedinosaurs", 48],
-  ["fotd", 48],
-  ["dinosaurs", 48],
-  ["crimbo22", 49],
-  ["gratitude", 49],
-  ["standard", 999],
-]);
-
-export const ORB_RESPONSES: string[] = [
-  "It is certain.",
-  "It is decidedly so.",
-  "Without a doubt.",
-  "Yes - definitely.",
-  "You may rely on it.",
-  "As I see it, yes.",
-  "Most likely.",
-  "Outlook good.",
-  "Yes.",
-  "Signs point to yes.",
-  "Reply hazy, try again.",
-  "Ask again later.",
-  "Better not tell you now.",
-  "Cannot predict now.",
-  "Concentrate and ask again.",
-  "Don't count on it.",
-  "My reply is no.",
-  "My sources say no.",
-  "Outlook not so good.",
-  "Very doubtful.",
-  "If CDM has a free moment, sure.",
-  "Why not?",
-  "How am I meant to know?",
-  "I guess???",
-  "...you realise that this is just a random choice from a list of strings, right?",
-  "I have literally no way to tell.",
-  "Ping Bobson, he probably knows",
-  "Check the wiki, answer's probably in there somewhere.",
-  "The wiki has the answer.",
-  "The wiki has the answer, but it's wrong.",
-  "I've not finished spading the answer to that question yet.",
-  "The devs know, go pester them instead of me.",
-  "INSUFFICIENT DATA FOR MEANINGFUL ANSWER",
-  "THERE IS AS YET INSUFFICIENT DATA FOR A MEANINGFUL ANSWER",
-];
-
 export const KILLMATCHER = /([A-Za-z0-9\-\_ ]+)\s+\(#\d+\)\s+defeated\D+(\d+)/;
 export const SKILLMATCHER = /([A-Za-z0-9\-\_ ]+)\s+\(#\d+\)\s+used the machine/;
 export const ITEMMATCHER = /\[\[([^\[\]]*)\]\]/g;
@@ -501,110 +357,3 @@ export const ROLEMAP: Map<string, string> = new Map([
   ["🚫", "512522219574919179"],
   ["✅", "754473984661258391"],
 ]);
-
-export const PROJECT_ALIASES: Map<string, string> = new Map([
-  ["garbo", "garbage-collector"],
-  ["freecandy", "freecandydotexe"],
-  ["garfjalen", "kol-scripting-resources"],
-  ["mafia", "kolmafia"],
-  ["oaf", "oaf-js"],
-]);
-
-export const PROJECT_CAPITALISATIONS: Map<string, string> = new Map([
-  ["tourguide", "TourGuide"],
-  ["chit", "ChIT"],
-  ["kol-scripting-resources", "KoL-Scripting-Resources"],
-  ["cagebot", "Cagebot"],
-  ["uberpvpoptimizer", "UberPvPOptimizer"],
-]);
-
-export const DREAD_BOSS_MAPPINGS: Map<string, string> = new Map([
-  ["werewolf", "Air Wolf"],
-  ["bugbear", "Falls-From-Sky"],
-  ["zombie", "Zombie HOA"],
-  ["ghost", "Mayor Ghost"],
-  ["vampire", "Drunkula"],
-  ["skeleton", "Unkillable Skeleton"],
-  ["xwerewolf", "~~Air Wolf~~"],
-  ["xbugbear", "~~Falls-From-Sky~~"],
-  ["xzombie", "~~Zombie HOA~~"],
-  ["xghost", "~~Mayor Ghost~~"],
-  ["xvampire", "~~Drunkula~~"],
-  ["xskeleton", "~~Unkillable Skeleton~~"],
-  ["unknown", "Boss unknown"],
-]);
-
-export enum ItemType {
-  Food = "Food",
-  Booze = "Booze",
-  Spleen = "Spleen Item",
-  Offhand = "Offhand",
-  SpecificFamiliarEquip = "Familiar Equipment (Specific)",
-  GenericFamiliarEquipment =  "Familiar Equipment (Generic)",
-  Unknown = "",
-}
-
-export const ITEM_SPADING_CALLS = [
-  {
-    url: (id: number) => [
-      "inv_equip.php",
-      {
-        action: "equip",
-        which: 2,
-        whichitem: id,
-      },
-    ],
-    visitMatch: /You can't equip an off-hand item while wielding a 2-handed weapon/,
-    type: ItemType.Offhand,
-  },
-  {
-    url: (id: number) => [
-      "inv_equip.php",
-      {
-        action: "equip",
-        which: 2,
-        whichitem: id,
-      },
-    ],
-    visitMatch: /Ghosts can't wear equipment/,
-    type: ItemType.GenericFamiliarEquipment,
-  },
-  {
-    url: (id: number) => [
-      "inv_eat.php",
-      {
-        which: 1,
-        whichitem: id,
-      },
-    ],
-    visitMatch: /You don't have the item you're trying to use/,
-    type: ItemType.Food,
-  },
-  {
-    url: (id: number) => [
-      "inv_booze.php",
-      {
-        which: 1,
-        whichitem: id,
-      },
-    ],
-    visitMatch: /You don't have the item you're trying to use/,
-    type: ItemType.Booze,
-  },
-  {
-    url: (id: number) => [
-      "inv_spleen.php",
-      {
-        which: 1,
-        whichitem: id,
-      },
-    ],
-    visitMatch: /You don't have the item you're trying to use/,
-    type: ItemType.Spleen,
-  },
-] as const;
-
-export enum SpadingFamiliars {
-  DEFAULT = 198,
-  GHOST = 282,
-}
