+++ src/index.ts
@@ -1,10 +1,8 @@
 import { DiscordClient } from "./discord";
 import { WikiSearcher } from "./wikisearch";
-import { attachMiscCommands, syncReminders } from "./misccommands";
-import { KOLClient } from "./kolclient";
-import { attachClanCommands, syncToDatabase } from "./raidlogs";
+import { KOLClient } from "./kol";
 import * as dotenv from "dotenv";
-import { attachKoLCommands } from "./kolcommands";
+import commands from "./commands";
 import { Pool } from "pg";
 import { migrate } from "postgres-migrations";
 
@@ -13,7 +11,7 @@ async function performSetup(): Promise<DiscordClient> {
   const kolClient = new KOLClient();
 
   console.log("Creating database client pool.");
-  const databaseClientPool = new Pool({
+  const databasePool = new Pool({
     connectionString: process.env.DATABASE_URL,
     ssl: {
       rejectUnauthorized: false,
@@ -24,10 +22,7 @@ async function performSetup(): Promise<DiscordClient> {
   const wikiSearcher = new WikiSearcher(kolClient);
 
   console.log("Migrating database.");
-  await migrate({ client: databaseClientPool }, "./migrations");
-
-  console.log("Syncing database.");
-  await syncToDatabase(databaseClientPool);
+  await migrate({ client: databasePool }, "./migrations");
 
   console.log("Downloading mafia data.");
   await wikiSearcher.downloadMafiaData();
@@ -36,20 +31,17 @@ async function performSetup(): Promise<DiscordClient> {
   console.log("Creating discord client.");
   const discordClient = new DiscordClient(wikiSearcher);
 
-  console.log("Attaching kol commands.");
-  attachKoLCommands(discordClient, kolClient, wikiSearcher);
+  const args = { discordClient, kolClient, wikiSearcher, databasePool };
 
-  console.log("Attaching clan commands.");
-  attachClanCommands(discordClient, kolClient, databaseClientPool);
-
-  console.log("Attaching misc commands.");
-  attachMiscCommands(discordClient, databaseClientPool);
+  for (const command of Object.values(commands)) {
+    command.attach(args);
+    if (command.sync) {
+      await command.sync(args);
+    }
+  }
 
   console.log("Attaching wiki commands.");
   discordClient.attachMetaBotCommands();
-
-  console.log("Syncing reminders.");
-  await syncReminders(databaseClientPool, discordClient.client());
 
   console.log("Registering slash commands.");
   await discordClient.registerSlashCommands();
