+++ src/kol.ts
@@ -4,7 +4,6 @@ import { cleanString, indent, toWikiLink } from "./utils";
 import { Mutex } from "async-mutex";
 import { DOMParser } from "xmldom";
 import { select } from "xpath";
-import { ItemType, ITEM_SPADING_CALLS } from "./constants";
 
 const clanActionMutex = new Mutex();
 const loginMutex = new Mutex();
@@ -93,14 +92,6 @@ type PlayerBasicData = {
   id: string;
   level: number;
   class: string;
-};
-
-type SpadedItem = {
-  id: number;
-  exists: boolean;
-  tradeable: boolean;
-  itemtype: ItemType;
-  additionalInfo?: string;
 };
 
 function sanitiseBlueText(blueText: string): string {
@@ -193,7 +184,7 @@ export class KOLClient {
     }
   }
 
-  private async tryRequestWithLogin(url: string, parameters: object) {
+  async tryRequestWithLogin(url: string, parameters: object) {
     const result = await this.makeCredentialedRequest(url, parameters);
     if (result) return result;
     await this.logIn();
@@ -535,61 +526,6 @@ export class KOLClient {
     } catch (error) {
       return undefined;
     }
-  }
-
-  async spadeItem(itemId: number): Promise<SpadedItem> {
-    let itemtype = ItemType.Unknown;
-    let additionalInfo = "";
-    const exists = !/Nopers/.test(
-      await this.tryRequestWithLogin("inv_equip.php", {
-        action: "equip",
-        which: 2,
-        whichitem: itemId,
-      })
-    );
-    const tradeable = !/That item cannot be sold or transferred/.test(
-      await this.tryRequestWithLogin("town_sellflea.php", {
-        whichitem: itemId,
-        sellprice: "",
-        selling: "Yep.",
-      })
-    );
-    if (exists) {
-      for (let property of ITEM_SPADING_CALLS) {
-        const { url, visitMatch, type } = property;
-        const page = (await this.tryRequestWithLogin(
-          ...(url(itemId) as [string, object])
-        )) as string;
-
-        const match = visitMatch.test(page);
-        if (match) {
-          itemtype = type;
-          break;
-        }
-      }
-    }
-    return { id: itemId, exists, tradeable, itemtype, additionalInfo };
-  }
-
-  async spadeFamiliar(famId: number): Promise<string> {
-    const page = await this.tryRequestWithLogin("desc_familiar.php", { which: famId });
-
-    if (page.includes("No familiar was found.")) return "none";
-
-    const name = /<font face=Arial,Helvetica><center><b>([^<]+)<\/b>/.exec(page)?.[1];
-    return name ?? "none";
-  }
-
-  async spadeSkill(skillId: number): Promise<boolean> {
-    const page = await this.tryRequestWithLogin("runskillz.php", {
-      action: "Skillz",
-      whichskill: skillId,
-      targetplayer: 1,
-      quantity: 1,
-    });
-
-    // If the skill doesn't exist on the dev server, the response ends with an exclamation mark
-    return page.includes("You don't have that skill.");
   }
 
   async getBasicDetailsForUser(name: string): Promise<PlayerBasicData> {
